# docs/api/evalsmith-public.openapi.yaml
openapi: 3.1.0
info:
  title: EvalSmith Public API
  version: 0.1.0
  summary: Gateway-facing API surface for EvalSmith trial and private deployments
  description: >
    First-pass public contract for EvalSmith. This spec focuses on the core
    integration paths external users need for auth, datasets, traces,
    experiments, annotation, and monitoring.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
...
paths:
...
  /api/v1/auth/access:
    get:
      tags: [auth]
      operationId: resolveAccess
      summary: Resolve access context for a session or API key
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Access context resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
...
  /api/v1/annotation/tasks:
    get:
      tags: [annotation]
      operationId: listAnnotationTasks
      summary: List annotation tasks
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Annotation task page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGenericEnvelope"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      __REDACTED_SECRET__: API Key
    sessionCookie:
      type: apiKey
      in: cookie
      name: evalsmith_session
  parameters:
    ProjectIdHeader:
      name: X-Project-ID
      in: header
      required: true
      schema:
        type: string
      description: Active project identifier
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        details:
          description: Optional error details
      required: [code, message]
    GenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data: {}
...
    PaginatedGenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: {}
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
                total_pages:
                  type: integer
...